requirements.txt
ffmpeg-python==0.2.0
requests==2.27.1